app/requirements.txt
mysql-connector-python==8.0.33
mariadb==1.1.12
cryptography==41.0.1
Flask==2.3.3
python.dotenv==1.0.0
bcrypt==4.3.0
email-validator==2.1.0
gunicorn==21.2.0